requirements.txt
mkdocs
mkdocs-material
numpy
Pillow
pre-commit
pytest
